decouple useNodePeers hook
